web/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
PyMuPDF>=1.25.0
python-docx>=1.1.0
diagram-forge @ file:../../
